% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{matched_study_logistics}
\alias{matched_study_logistics}
\title{Logistics for Matched Study}
\usage{
matched_study_logistics(
  n_pairs,
  m = 1,
  pair_identification_rate,
  pair_rejection_rate,
  working_days_month
)
}
\arguments{
\item{n_pairs}{Number of required pairs}

\item{m}{Number of controls per case}

\item{pair_identification_rate}{Rate of pair identification per day}

\item{pair_rejection_rate}{Expected rejection rate for pairs}

\item{working_days_month}{Number of working days per month}
}
\value{
A list with study logistics calculations
}
\description{
Logistics for Matched Study
}
